move editable.directive
